refactor: 패키지 구조 도메인별로 구분하도록 개선
