+++ src/main/java/org/sopt/common/code/ErrorCode.java
@@ -1,0 +1,12 @@
+package org.sopt.common.code;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCode {
+
+    String name();
+
+    HttpStatus getStatus();
+
+    String getMessage();
+}
