+++ src/main/java/org/sopt/common/code/SuccessCode.java
@@ -1,0 +1,8 @@
+package org.sopt.common.code;
+
+public interface SuccessCode {
+
+    String name();
+
+    String getMessage();
+}
