+++ src/main/java/org/sopt/common/exception/BusinessException.java
@@ -1,0 +1,16 @@
+package org.sopt.common.exception;
+
+import org.sopt.common.code.ErrorCode;
+
+public class BusinessException extends RuntimeException {
+    private final ErrorCode errorCode;
+
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+
+    public ErrorCode getErrorCode() {
+        return errorCode;
+    }
+}
