--- src/main/java/org/sopt/controller/dto/request/UpdatePostRequest.java
@@ -1,13 +1,0 @@
-package org.sopt.controller.dto.request;
-
-import org.sopt.service.vo.UpdatePostCommand;
-
-public record UpdatePostRequest(
-        String newTitle,
-        String newContent
-) {
-
-    public UpdatePostCommand toCommand() {
-        return new UpdatePostCommand(newTitle, newContent);
-    }
-}
