--- src/main/java/org/sopt/enums/BoardType.java
@@ -1,7 +1,0 @@
-package org.sopt.enums;
-
-public enum BoardType {
-    FREE, //자유
-    HOT, //핫게
-    SECRET, //비밀
-}
