--- src/main/java/org/sopt/exception/BusinessException.java
@@ -1,16 +1,0 @@
-package org.sopt.exception;
-
-import org.sopt.enums.ErrorMessage;
-
-public class BusinessException extends RuntimeException {
-    private final ErrorMessage errorMessage;
-
-    public BusinessException(ErrorMessage errorMessage) {
-        super(errorMessage.getClientMessage());
-        this.errorMessage = errorMessage;
-    }
-
-    public ErrorMessage getErrorMessage() {
-        return errorMessage;
-    }
-}
