--- src/main/java/org/sopt/exception/PostNotFoundException.java
@@ -1,9 +1,0 @@
-package org.sopt.exception;
-
-import org.sopt.enums.ErrorMessage;
-
-public class PostNotFoundException extends BusinessException {
-    public PostNotFoundException() {
-        super(ErrorMessage.POST_NOT_FOUND);
-    }
-}
