+++ src/main/java/org/sopt/post/controller/dto/response/GetAllPostsResponse.java
@@ -1,0 +1,14 @@
+package org.sopt.post.controller.dto.response;
+
+import java.util.List;
+import org.sopt.post.domain.Post;
+
+public record GetAllPostsResponse(
+        List<PostResponse> posts
+) {
+    public static GetAllPostsResponse of(List<Post> posts) {
+        return new GetAllPostsResponse(
+                posts.stream().map(PostResponse::from).toList()
+        );
+    }
+}
