+++ src/main/java/org/sopt/post/enums/BoardType.java
@@ -1,0 +1,7 @@
+package org.sopt.post.enums;
+
+public enum BoardType {
+    FREE, //자유
+    HOT, //핫게
+    SECRET, //비밀
+}
