+++ src/main/java/org/sopt/post/exception/PostNotFoundException.java
@@ -1,0 +1,11 @@
+package org.sopt.post.exception;
+
+
+import org.sopt.common.exception.BusinessException;
+import org.sopt.post.code.PostErrorCode;
+
+public class PostNotFoundException extends BusinessException {
+    public PostNotFoundException() {
+        super(PostErrorCode.POST_NOT_FOUND);
+    }
+}
