+++ src/main/java/org/sopt/post/service/vo/UpdatePostCommand.java
@@ -1,0 +1,7 @@
+package org.sopt.post.service.vo;
+
+public record UpdatePostCommand(
+        String newTitle,
+        String newContent
+) {
+}
