--- src/main/java/org/sopt/repository/PostRepository.java
@@ -1,23 +1,0 @@
-package org.sopt.repository;
-
-import java.util.List;
-import java.util.Optional;
-import org.sopt.domain.Post;
-import org.sopt.enums.BoardType;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository {
-
-    Post save(Post post);
-
-    List<Post> findAll(int page, int size);
-
-    Optional<Post> findById(Long id);
-
-    void deleteById(Long id);
-
-    Long generateId();
-
-    List<Post> findAllByBoardType(BoardType boardType, int page, int size);
-}
