--- src/main/java/org/sopt/service/vo/CreatePostCommand.java
@@ -1,11 +1,0 @@
-package org.sopt.service.vo;
-
-import org.sopt.enums.BoardType;
-
-public record CreatePostCommand(
-        String title,
-        String content,
-        String author,
-        BoardType boardType
-) {
-}
