--- src/main/java/org/sopt/service/vo/UpdatePostCommand.java
@@ -1,7 +1,0 @@
-package org.sopt.service.vo;
-
-public record UpdatePostCommand(
-        String newTitle,
-        String newContent
-) {
-}
